[Refactor] new folder `src/base/common/structures` to hold all kinds of data structures
