+++ src/base/browser/basic/contextMenu/contextMenu.ts
@@ -3,7 +3,7 @@ import { addDisposableListener, DomUtility, EventType } from "src/base/browser/b
 import { FastElement } from "src/base/browser/basic/fastElement";
 import { AnchorAbstractPosition, AnchorMode, calcViewPositionAlongAxis, IAnchorBox } from "src/base/browser/basic/view";
 import { Disposable, DisposableManager, IDisposable } from "src/base/common/dispose";
-import { Range } from "src/base/common/range";
+import { Range } from "src/base/common/structures/range";
 import { IDomBox, IPosition } from "src/base/common/util/size";
 
 export interface IAnchor {
