+++ src/base/browser/secondary/listView/list.ts
@@ -1,7 +1,7 @@
 import { IListView, IViewItemChangeEvent } from "src/base/browser/secondary/listView/listView";
 import { IListWidget } from "src/base/browser/secondary/listWidget/listWidget";
 import { Register } from "src/base/common/event";
-import { IRange } from "src/base/common/range";
+import { IRange } from "src/base/common/structures/range";
 
 /**
  * Common interfaces shares between {@link IListView} and {@link IListWidget}.
