+++ src/base/common/structures/linkedList.ts
