+++ src/base/common/structures/map.ts
