+++ src/base/common/structures/range.ts
