+++ src/base/common/structures/ternarySearchTree.ts
