+++ src/base/common/util/array.ts
@@ -1,6 +1,4 @@
-import { IDisposable } from "src/base/common/dispose";
-import { IIterable } from "src/base/common/util/iterable";
-import { CompareFn, Mutable, NonUndefined } from "src/base/common/util/type";
+import { NonUndefined } from "src/base/common/util/type";
 
 /**
  * @namespace Array A series of helper functions that relates to array.
@@ -393,472 +391,5 @@ export namespace Arrays {
         }
 
         return undefined;
-    }
-}
-
-/**
- * Interface for {@link IDeque}.
- */
-export interface IDeque<T> extends IIterable<T>, IDisposable {
-    size(): number;
-    empty(): boolean;
-    
-    at(index: number): T;
-    front(): T;
-    back(): T;
-    
-    pushBack(element: T): void;
-    pushFront(element: T): void;
-    popBack(): T;
-    popFront(): T;
-    
-    insert(index: number, element: T): void;
-    remove(index: number): T;
-    replace(index: number, element: T): T;
-    swap(first: number, second: number): void;
-    reverse(): void;
-    
-    extendFront(other: Deque<T>): void;
-    extendBack(other: Deque<T>): void;
-    clear(): number;
-}
-
-export class Deque<T> implements IDeque<T> {
-
-    // [field]
-
-    private _arr: T[];
-    
-    // [constructor]
-
-    constructor(elements: T[] = []) {
-        this._arr = [];
-        for (const element of elements) {
-            this._arr.push(element);
-        }
-    }
-
-    // [public methods]
-
-    public size(): number {
-        return this._arr.length;
-    }
-
-    public empty(): boolean {
-        return !this._arr.length;
-    }
-
-    public at(index: number): T {
-        if (index < 0 || index >= this.size()) {
-            throw new Error(`Invalid index when getting elements in deque at ${index}.`);
-        }
-        return this._arr[index]!;
-    }
-
-    public front(): T {
-        return this.at(0);
-    }
-
-    public back(): T {
-        return this.at(this.size() - 1);
-    }
-
-    public pushBack(element: T): void {
-        this._arr.splice(this.size(), 0, element);
-    }
-
-    public pushFront(element: T): void {
-        this._arr.splice(0, 0, element);
-    }
-
-    public popBack(): T {
-        if (this.empty()) {
-            throw new Error('Cannot pop back from deque because it is empty.');
-        }
-        return this._arr.pop()!;
-    }
-
-    public popFront(): T {
-        if (this.empty()) {
-            throw new Error('Cannot pop front from deque because it is empty.');
-        }
-        return this._arr.splice(0, 1)[0]!;
-    }
-
-    public insert(index: number, element: T): void {
-        if (index < 0 || index > this.size()) {
-            throw new Error(`Invalid index when inserting elements in deque at ${index}.`);
-        }
-        this._arr.splice(index, 0, element);
-    }
-
-    public remove(index: number): T {
-        const removed = this._arr.splice(index, 1);
-        if (!removed.length) {
-            throw new Error(`Cannot remove elements from deque at invalid index ${index}.`);
-        }
-        return removed[0]!;
-    }
-
-    public replace(index: number, element: T): T {
-        const old = this.at(index);
-        this._arr[index] = element;
-        return old;
-    }
-
-    public swap(first: number, second: number): void {
-        const firstElement = this.at(first);
-        this._arr[first] = this.at(second);
-        this._arr[second] = firstElement;
-    }
-
-    public reverse(): void {
-        this._arr.reverse();
-    }
-
-    public extendFront(other: Deque<T>): void {
-        for (const element of other) {
-            this.pushFront(element);
-        }
-    }
-
-    public extendBack(other: Deque<T>): void {
-        for (const element of other) {
-            this.pushBack(element);
-        }
-    }
-
-    public clear(): number {
-        const removed = this.size();
-        this._arr = [];
-        return removed;
-    }
-
-    public dispose(): void {
-        this.clear();
-    }
-
-    *[Symbol.iterator](): Iterator<T> {
-		let idx = 0;
-        while (idx < this.size()) {
-			yield this._arr[idx]!;
-			idx += 1;
-		}
-	}
-}
-
-/**
- * Interface for {@link Stack}.
- */
-export interface IStack<T> extends IIterable<T>, IDisposable {
-    size(): number;
-    empty(): boolean;
-    top(): T;
-    push(element: T): void;
-    pop(): T;
-    clear(): void;
-}
-
-export class Stack<T> implements IStack<T> {
-
-    private readonly _deque: Deque<T>;
-
-    constructor(elements: T[] = []) {
-        this._deque = new Deque(elements);
-    }
-
-    public size(): number {
-        return this._deque.size();
-    }
-
-    public empty(): boolean {
-        return this._deque.empty();
-    }
-
-    public top(): T {
-        return this._deque.front();
-    }
-
-    public push(element: T): void {
-        this._deque.pushFront(element);
-    }
-
-    public pop(): T {
-        return this._deque.popFront();
-    }
-
-    public clear(): void {
-        this._deque.clear();
-    }
-
-    public dispose(): void {
-        this.clear();
-    }
-
-    *[Symbol.iterator](): Iterator<T> {
-		let idx = 0;
-        while (idx < this.size()) {
-			yield this._deque.at(idx);
-			idx += 1;
-		}
-	}
-}
-
-/**
- * Interface only for {@link Queue}.
- */
-export interface IQueue<T> extends IIterable<T>, IDisposable {
-    size(): number;
-    empty(): boolean;
-    front(): T;
-    back(): T;
-    pushBack(element: T): void;
-    popFront(): T;
-    clear(): void;
-}
-
-export class Queue<T> implements IQueue<T> {
-
-    private readonly _deque: Deque<T>;
-
-    constructor(elements: T[] = []) {
-        this._deque = new Deque(elements);
-    }
-
-    public size(): number {
-        return this._deque.size();
-    }
-
-    public empty(): boolean {
-        return this._deque.empty();
-    }
-
-    public front(): T {
-        return this._deque.front();
-    }
-
-    public back(): T {
-        return this._deque.back();
-    }
-
-    public pushBack(element: T): void {
-        this._deque.pushBack(element);
-    }
-
-    public popFront(): T {
-        return this._deque.popFront();
-    }
-
-    public clear(): void {
-        this._deque.clear();
-    }
-
-    public dispose(): void {
-        this.clear();
-    }
-
-    *[Symbol.iterator](): Iterator<T> {
-		let idx = 0;
-        while (idx < this.size()) {
-			yield this._deque.at(idx);
-			idx += 1;
-		}
-	}
-}
-
-export interface IPriorityQueue<T> extends IIterable<T>, IDisposable {
-	
-	/**
-     * @description Adds an element to the queue.
-     * @param element - The element to add.
-     */
-	enqueue(element: T): void;
-		
-	/**
-	 * @description Removes and returns the highest-priority element.
-	 * @returns The highest-priority element or undefined if the queue is empty.
-	 */
-	dequeue(): T | undefined;
-
-    /**
-     * @description Removes the specified element from the priority queue. This 
-     * method finds all occurrences of the given element in the heap, replaces 
-     * them with the last element in the heap, and then sifts up or down as 
-     * necessary to maintain the heap properties. 
-     * @param element - The element to be removed from the priority queue.
-     * @complexity O(n) - where n is the number of elements in the heap.
-     * @throws If the element is not found.
-     */
-    remove(element: T): void;
-
-	/**
-	 * @description Returns the highest-priority element without removing it.
-	 * @returns The highest-priority element or undefined if the queue is empty.
-	 */
-	peek(): T | undefined;
-
-	/**
-	 * @description Returns the number of elements in the queue.
-	 * @returns The number of elements in the queue.
-	 */
-	size(): number;
-
-	/**
-	 * @description Checks if the queue is empty.
-	 * @returns True if the queue is empty, false otherwise.
-	 */
-	isEmpty(): boolean;
-}
-
-export class PriorityQueue<T> implements IPriorityQueue<T> {
-	
-	// [fields]
-
-    private _heap: T[] = [];
-    private _count: number = 0;
-    private readonly __comparator: CompareFn<T>;
-
-	// [constructor]
-
-    constructor(comparator: CompareFn<T>) {
-        this.__comparator = comparator;
-    }
-
-	// [public methods]
-    
-    public enqueue(element: T): void {
-        this._heap[this._count] = element;
-        this._count++;
-        this.__bubbleUp();
-    }
-    
-    public dequeue(): T | undefined {
-        if (this.isEmpty()) {
-            return undefined;
-        }
-        
-        const item = this._heap[0];
-        this._heap[0] = this._heap[this._count - 1]!;
-        this._heap.pop();
-        
-        this._count--;
-        this.__sinkDown();
-
-        return item;
-    }
-
-    public remove(element: T): void {
-        const index = this._heap.findIndex(e => e === element);
-    
-        if (index === -1) {
-            return;
-        }
-    
-        this._heap[index] = this._heap[this._count - 1]!;
-        this._heap.pop();
-        this._count--;
-    
-        // If the element was the last or only element, no reordering is necessary.
-        if (index === this._count) {
-            return;
-        }
-    
-        // Decide whether to bubble up or sink down
-        const parentIdx = Math.floor((index - 1) / 2);
-        if (index > 0 && this.__comparator(this._heap[index]!, this._heap[parentIdx]!) < 0) {
-            this.__bubbleUp(index);
-        } else {
-            this.__sinkDown(index);
-        }
-    }
-    
-    public peek(): T | undefined {
-        if (this.isEmpty()) {
-            return undefined;
-        }
-        return this._heap[0];
-    }
-    
-    public size(): number {
-        return this._count;
-    }
-    
-    public isEmpty(): boolean {
-        return this._heap.length === 0;
-    }
-
-    public dispose(): void {
-        this._heap = [];
-        this._count = 0;
-    }
-
-	*[Symbol.iterator](): Iterator<T> {
-        const copy = [...this._heap];
-        const copyCount = this._count;
-        const copyComparator = this.__comparator;
-
-        const copyQueue = new PriorityQueue<T>(copyComparator);
-        copyQueue._heap = copy;
-        copyQueue._count = copyCount;
-        
-        while (!copyQueue.isEmpty()) {
-            yield copyQueue.dequeue()!;
-        }
-    }
-    
-    // [private helper methods]
-
-    private __bubbleUp(index: number = this._count - 1): void {
-        while (index > 0) {
-            const item = this._heap[index]!;
-            const parentIdx = Math.floor((index - 1) / 2);
-            const parent = this._heap[parentIdx]!;
-            
-            if (this.__comparator(item, parent) >= 0) {
-                break;
-            }
-            
-            this._heap[index] = parent;
-            this._heap[parentIdx] = item;
-            index = parentIdx;
-        }
-    }
-
-    private __sinkDown(index: number = 0): void {
-        const length = this.size();
-        const element = this._heap[0]!;
-
-        while (true) {
-            const leftChildIdx = 2 * index + 1;
-            const rightChildIdx = 2 * index + 2;
-            
-            let swapIdx: number | null = null;
-
-            if (leftChildIdx < length) {
-                const leftChild = this._heap[leftChildIdx]!;
-                if (this.__comparator(leftChild, element) < 0) {
-                    swapIdx = leftChildIdx;
-                }
-            }
-            
-            if (rightChildIdx < length) {
-                const leftChild = this._heap[leftChildIdx]!;
-                const rightChild = this._heap[rightChildIdx]!;
-                if (
-                    (swapIdx === null && this.__comparator(rightChild, element) < 0) || 
-                    (swapIdx !== null && this.__comparator(rightChild, leftChild) < 0)
-                ) {
-                    swapIdx = rightChildIdx;
-                }
-            }
-
-            if (swapIdx === null) {
-                break;
-            }
-
-            this._heap[index] = this._heap[swapIdx]!;
-            this._heap[swapIdx] = element;
-            index = swapIdx;
-        }
     }
 }
