+++ src/platform/logger/common/abstractLoggerService.ts
@@ -1,7 +1,7 @@
 import { Disposable } from "src/base/common/dispose";
 import { URI } from "src/base/common/file/uri";
 import { DEFAULT_LOG_LEVEL, ILogger, ILoggerOpts, LogLevel } from "src/base/common/logger";
-import { ResourceMap } from "src/base/common/util/map";
+import { ResourceMap } from "src/base/common/structures/map";
 import { IService, createService } from "src/platform/instantiation/common/decorator";
 
 export const ILoggerService = createService<ILoggerService>('logger-service');
