+++ test/base/common/structures/ternarySearchTree.test.ts
@@ -1,6 +1,6 @@
 import * as assert from 'assert';
 import { URI } from 'src/base/common/file/uri';
-import { CreateTernarySearchTree, ITernarySearchTree, StringIterator, TernarySearchTree, TernarySearchTreeNode, UriIterator } from 'src/base/common/util/ternarySearchTree';
+import { CreateTernarySearchTree, ITernarySearchTree, StringIterator, TernarySearchTree, TernarySearchTreeNode, UriIterator } from 'src/base/common/structures/ternarySearchTree';
 
 suite('ternarySearchTree-test', () => {
